// src/app/models/tv-shows.response.model.ts
export interface TvShow {
    adult: boolean;
    backdrop_path: string;
    genre_ids: string[];
    id: number;
    origin_country: string[],
    original_language: string;
    original_name: string;
    overview: string;
    popularity: number;
    poster_path: string;
    first_air_date: string;
    name: string;
    vote_average: number;
    vote_count: number,
  }
  
export interface TvShowsResponse {
    page: number;
    results: TvShow[];
    total_pages: number;
    total_results: number;
}
  